refac: 프로필 이미지 업로드 부분 클래스 분리
